Use Hash#fetch consistently in unsigned_download_url
This should have no effect on behaviour. I had to change some specs
that were checking for side effects on the options hash. This hash
is not passed on to any other methods, so it should be safe not to
delete keys from it. The assertions on options only make for fragile
specs.