Fix the overwritten initializer for hash
I ran into this in a Rails project. I wanted to call `#slice` on the Response.
I was using the Admin API and making a single response call.
I've replaced the public_id with 'foo'.

```
response = Cloudinary::Api.resource "foo"
=> { "public_id" => 'foo', ...  }

response.is_a? Hash
=> true

response.respond_to? :slice
=> true

response.slice 'public_id'
=> ArgumentError: wrong number of arguments (given 0, expected 1)
  from /Users/mattsullivan/.rvm/gems/ruby-2.3.4/gems/cloudinary-1.8.1/lib/cloudinary/api.rb:16:in `initialize'
  from /Users/mattsullivan/.rvm/gems/ruby-2.3.4/gems/activesupport-4.2.10/lib/active_support/core_ext/hash/slice.rb:23:in `new'
  from /Users/mattsullivan/.rvm/gems/ruby-2.3.4/gems/activesupport-4.2.10/lib/active_support/core_ext/hash/slice.rb:23:in `slice'
```
The response inherits from Hash. By overwriting the initializer
we break the parent class when used in a Rails project.
SEE #slice https://apidock.com/rails/Hash/slice.
It tries to re-instantiate a new hash, but can't because it doesn't have a response, which the new #initializer requires.

The goal here is to maintain continuity with the Parent class(Hash).
I'm guessing this gem is used mainly for Rails. I added the rails hash
extensions to the response so if a user uses this lib in a Rails project the
Response Hash will behave as expected. If a user uses it in a vanilla ruby script
etc. it should behave as expected.

Caveat: We don't have to require active_support's hash ext. in the api.rb.
The main change was to allow the response to be nil. This might be more accurate
since if a user is say using Sinatra or something they wouldn't expect hash to
have those extra methods.