Fixed direct references to ::CarrierWave with `defined?` guards.
The migrator has zero tests to work against but this fixes our rails
app that does not use carrierwave. I'm not thrilled with the guard on
the require, but guarding the entire class inside of
carrier_wave/storage seems worse. If cloudinary/carrier_wave.rb really
is supposed to be the only entrypoint then it seems fine.

This project needs (more) tests.

This project needs to check against all the envs that are supposedly supported.