Avoid early loading of ActionView::Base
The injection of the CloudinaryHelper module can happen lazily when
ActionView::Base is loaded by some other code needing it, and there is
no reason to trigger the loading early in the railtie.

A consequence of this change is that you can avoid the loading of
ActionView::Base altogether in some cases, e.g. if running a specific
test that does not need it, or when opening a rails console.